requirements.txt
certifi==2017.4.17
chardet==3.0.3
Django==1.11.2
django-filter==1.0.4
djangorestframework==3.6.3
honcho==1.0.1
idna==2.5
Markdown==2.6.8
psycopg2==2.7.1
pytz==2017.2
requests==2.17.3
ujson==1.35
urllib3==1.21.1
